Remote correctly sending a name to the server
A few things to note:
I'm using asyncio in a synchronous manner (in our
case it's no big deal but in principle I would like to avoid it) due to
asyncio.run complaining about different event loops (looks like more
loops are created). For static use, I'm directly referring to the event
loop (asyncio.get_event_loop).

Note how the writer should be the non-remote player (eg. white, if black
is the remote), but the code is structured so that the Remote class
manages all the input-output with the server, hence it has to speak for the other player
(eg. send its name to the Remote). This may or may not change in the
future, for a better separation of jobs.

When two players are connected, they both successfully send a name, but
everything instantly crashes, as the actual game management has to be
refined.