Better remote (kinda... working?)
This commit should have been divided into multiples but I don't have the time.
Important changes:
Client now have different methods for reading and sending data. Often times we only need to read or send, so it simply makes more sense to me.

msg_unpack now reads the integer and decodes only the specified amount of bytes. The buffer is not guaranteed to be cleaned up after a previous read, hence we have to carefully select the right data.

Managed the first action (None).

Added some reads here and there to consume the server very verbose messages. Eg. the server will echo the current position after each move to the player who did the move. We manage the state internally so I simply consume the input and wait for the next one (which is supposed to be the updated state from the other player's move).